Design a Business Card
Every business knows the importance of business cards. It is termed the most important branding material that allows people to know about your point of contact. Business cards showcase the identity and build a reputation for the business. it is a very basic and demanding project in graphic design. You will learn a lot in this project.